Extend AVAudioPlayerNode to be able to play OGG Vorbis files
This is deliberately simple at the moment, it simply loads the entire file into a single buffer and queues it.

Experimenting with smaller buffers worked, but had lots of clicks and ticks, so I'm not sure what the ideal approach there is yet. At least by having this method, we can experiment in one place.