Add a start handler for overlapping handling
BUG: When samples are 30s long, isOverlapping is true, and the sampleGap is 20s

This is intended to create a 10s overlap between samples, but the side-effect of always scheduling overlapping samples together means we infinitely loop scheduling each overlapped sample. The loop can only be broken by having a sample delay be longer than a sample length, which is not a constraint of an element playlist.

FIX: Schedule a zero-length buffer in the player before the file's buffer; its completion handler is thus a "start handler" telling us exactly when a buffer starts being played.

Move the delayBeforeNext < 0 handling into this start handler, so they're asynchronously scheduled in a reasonable manner.

Also revamp is the isOverlapping handling, to avoid subtracting and re-adding the sample length, instead just handle it as a separate branch of the `if` and use the delay directly.