Switch from a .framework to an .xcframework
Hit this error when building for the simulator:
> building for iOS Simulator, but linking in object file (…) built for Mac Catalyst

Turns out my guess that Catalyst and Simulator shared a library file
was not correct, and there's no decently proper way to build a
framework correctly for all three.

But an .xcframework works fine, it's effectively just a copy of all
three in one hierarchy, so build one of those instead.