Refactor out isPlaying into a computed property
It's now actually the combination of whether we're playing something or downloading something, which is what it should be.

Remembering that only the status updates are "will change", and that updating the playing and downloading lists is inherent, we can now simplify the stop() behavior a lot and remove the special-cased cancel behavior entirely (other than not printing an error).

The only gotcha is still sending the extra progress update since that's a current value publisher rather than a willChange.