Completely redo the one shots list
Building on SectionedList and Grid we can now make a PlaylistsBySoundset that can operate on an NSFetchRequest<Playlist>, section by the soundset object, add titles, and then use a Grid of PlaylistCell to layout the playlists.

Use this for the isSearching case, and to make everything clean, for the other case actually just use the PlaylistGrid we made before which also takes a fetch request, and also uses a Grid of PlaylsitCell just like the above.

This then makes the OneShots view itself super clean since it's just the UI components, and calling out to get fetch requests - rather than use a controller just have a static function on Playlist instead which simplifies things a lot!